reorganized the DTOs into mirror packages
