+++ src/main/java/org/braekpo1nt/mctmanager/config/dto/net/kyori/adventure/sound/SoundDTO.java
@@ -1,8 +1,9 @@
-package org.braekpo1nt.mctmanager.config.dto;
+package org.braekpo1nt.mctmanager.config.dto.net.kyori.adventure.sound;
 
 import com.google.common.base.Preconditions;
 import lombok.Data;
 import net.kyori.adventure.sound.Sound;
+import org.braekpo1nt.mctmanager.config.dto.org.bukkit.NamespacedKeyDTO;
 import org.braekpo1nt.mctmanager.config.validation.Validatable;
 import org.braekpo1nt.mctmanager.config.validation.Validator;
