+++ src/main/java/org/braekpo1nt/mctmanager/config/dto/org/bukkit/enchantments/EnchantmentDTO.java
@@ -1,14 +1,12 @@
-package org.braekpo1nt.mctmanager.config.dto.enchantments;
+package org.braekpo1nt.mctmanager.config.dto.org.bukkit.enchantments;
 
 import com.google.common.base.Preconditions;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.braekpo1nt.mctmanager.config.dto.NamespacedKeyDTO;
-import org.braekpo1nt.mctmanager.config.validation.ConfigInvalidException;
+import org.braekpo1nt.mctmanager.config.dto.org.bukkit.NamespacedKeyDTO;
 import org.braekpo1nt.mctmanager.config.validation.Validatable;
 import org.braekpo1nt.mctmanager.config.validation.Validator;
-import org.bukkit.NamespacedKey;
 import org.bukkit.enchantments.Enchantment;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
