+++ src/main/java/org/braekpo1nt/mctmanager/config/dto/org/bukkit/inventory/InventoryDTO.java
@@ -1,4 +1,4 @@
-package org.braekpo1nt.mctmanager.config.dto.inventory;
+package org.braekpo1nt.mctmanager.config.dto.org.bukkit.inventory;
 
 import com.google.common.base.Preconditions;
 import lombok.Getter;
