+++ src/main/java/org/braekpo1nt/mctmanager/config/dto/org/bukkit/inventory/meta/PotionMetaDTO.java
@@ -1,4 +1,4 @@
-package org.braekpo1nt.mctmanager.config.dto.inventory.meta;
+package org.braekpo1nt.mctmanager.config.dto.org.bukkit.inventory.meta;
 
 import com.google.common.base.Preconditions;
 import lombok.Data;
