+++ src/main/java/org/braekpo1nt/mctmanager/games/game/capturetheflag/config/LoadoutDTO.java
@@ -5,8 +5,8 @@ import com.google.gson.JsonElement;
 import lombok.Getter;
 import org.braekpo1nt.mctmanager.games.game.capturetheflag.Loadout;
 import org.braekpo1nt.mctmanager.config.ConfigUtils;
-import org.braekpo1nt.mctmanager.config.dto.inventory.PlayerInventoryDTO;
-import org.braekpo1nt.mctmanager.config.dto.inventory.meta.ItemMetaDTO;
+import org.braekpo1nt.mctmanager.config.dto.org.bukkit.inventory.PlayerInventoryDTO;
+import org.braekpo1nt.mctmanager.config.dto.org.bukkit.inventory.meta.ItemMetaDTO;
 import org.bukkit.Material;
 import org.bukkit.inventory.ItemStack;
 import org.jetbrains.annotations.NotNull;
