+++ src/main/java/org/braekpo1nt/mctmanager/games/game/footrace/config/FootRaceConfig.java
@@ -1,8 +1,8 @@
 package org.braekpo1nt.mctmanager.games.game.footrace.config;
 
 import com.google.gson.JsonElement;
-import org.braekpo1nt.mctmanager.config.dto.BoundingBoxDTO;
-import org.braekpo1nt.mctmanager.config.dto.LocationDTO;
+import org.braekpo1nt.mctmanager.config.dto.org.bukkit.util.BoundingBoxDTO;
+import org.braekpo1nt.mctmanager.config.dto.org.bukkit.LocationDTO;
 
 record FootRaceConfig(String version, String world, LocationDTO startingLocation, BoundingBoxDTO finishLine, BoundingBoxDTO spectatorArea, BoundingBoxDTO glassBarrier, Scores scores, Durations durations, JsonElement description) {
