+++ src/main/java/org/braekpo1nt/mctmanager/games/game/spleef/config/SpleefConfigDTO.java
@@ -7,9 +7,9 @@ import com.google.gson.JsonSyntaxException;
 import lombok.Data;
 import net.kyori.adventure.text.serializer.gson.GsonComponentSerializer;
 import org.braekpo1nt.mctmanager.Main;
-import org.braekpo1nt.mctmanager.config.dto.BoundingBoxDTO;
-import org.braekpo1nt.mctmanager.config.dto.NamespacedKeyDTO;
-import org.braekpo1nt.mctmanager.config.dto.inventory.ItemStackDTO;
+import org.braekpo1nt.mctmanager.config.dto.org.bukkit.util.BoundingBoxDTO;
+import org.braekpo1nt.mctmanager.config.dto.org.bukkit.NamespacedKeyDTO;
+import org.braekpo1nt.mctmanager.config.dto.org.bukkit.inventory.ItemStackDTO;
 import org.braekpo1nt.mctmanager.config.validation.ConfigInvalidException;
 import org.braekpo1nt.mctmanager.config.validation.Validatable;
 import org.braekpo1nt.mctmanager.config.validation.Validator;
